+++ src/main/java/subway/api/StationController.java
@@ -1,7 +1,10 @@
-package subway;
+package subway.api;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import subway.station.dto.StationRequest;
+import subway.station.dto.StationResponse;
+import subway.station.service.StationService;
 
 import java.net.URI;
 import java.util.List;
